Add files via upload
I wanted to use numerous header sizes, so I put two hashtags before my name to make it a header 2 size and four hashtags before "Introduction" to make it a header 4 size. To italicize the sources of the data I am using for my thesis, I surrounded the names of the studies by asterisks. To make the components of my thesis a list, I put a single asterisk and a space before the names of the policies. To embed a photo of myself, I copied the path to an image I uploaded to my repository on github and followed the code guidelines on https://www.rstudio.com/wp-content/uploads/2015/03/rmarkdown-reference.pdf.