Update README.md
I wanted to use numerous header sizes, so I put two hashtags before my name to make it a header 2 size and four hashtags before "Introduction" to make it a header 4 size. To italicize the sources of the data I am using for my thesis, I surrounded the names of the studies by asterisks. To make the components of my thesis a list, I put a single asterisk and a space before the names of the policies. Originally, I did not include an empty line after the list, so the text for the list flowed into the rest of the sentence rather than there being a clear break between the list and the rest of the text. I fixed this by adding an empty line after the last part of the list. To embed a photo of myself, I initially copied the path to an image I uploaded to my repository on github and followed the code guidelines on https://www.rstudio.com/wp-content/uploads/2015/03/rmarkdown-reference.pdf. It took a few tries to realize that I needed to upload the image to my "hw01" repository rather than my "myrepo" repository. I resized the image, uploaded it to my "hw01" repo, copied the new file path, and replaced the original path with the new one.